# python/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RAG Orchestrator API",
    "description": "API for uploading and searching multimedia files using the RAG Orchestrator.",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8001"
    }
  ],
  "paths": {
    "/upload": {
      "post": {
        "summary": "Upload a multimedia file",
        "operationId": "upload_multimedia_file",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "post": {
        "summary": "Search for multimedia content",
        "operationId": "search_multimedia_content",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProcessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "doc_id": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "SearchRequest": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string"
          },
          "file_types": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "top_k": {
            "type": "integer"
          }
        },
        "required": ["query"]
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            }
          },
          "total_count": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "SearchResult": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number"
          },
          "filename": {
            "type": "string"
          },
          "file_type": {
            "type": "string"
          },
          "content_type": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          }
        }
      }
    }
  }
}